Create card collection page to DRY up flashcard pages
