--- src/components/SpellingCheck/SpellingCheck.js
@@ -1,1 +1,0 @@
-import { h, Component } from 'preact';
